refactor: split modbus helper responsibilities
Extract call dispatch, frame logging, framer detection and client-close
helpers from the monolithic modbus_helpers.py into focused modules under
modbus/ (call.py, frame_logging.py, framer.py, client_close.py).
Move read-planning utilities (group_reads, chunk_register_range,
chunk_register_values) into registers/read_planner.py where they belong.
Update all callers to import from the new focused modules; modbus_helpers
is kept as a backward-compatible re-export shim.
