+++ custom_components/thessla_green_modbus/_coordinator_read_bits.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from .coordinator.retry import _PermanentModbusError
 from .modbus_exceptions import ConnectionException, ModbusException
-from .modbus_helpers import chunk_register_range
+from .registers.read_planner import chunk_register_range
 
 
 async def read_coil_registers_optimized(owner: Any) -> dict[str, Any]:
