+++ custom_components/thessla_green_modbus/coordinator/coordinator.py
@@ -82,7 +82,8 @@ from ..coordinator_state import (
 from ..coordinator_state import normalize_serial_settings as _normalize_serial_settings_impl
 from ..coordinator_state import resolve_effective_batch as _resolve_effective_batch_impl
 from ..errors import CannotConnect
-from ..modbus_helpers import group_reads
+from ..modbus_exceptions import ConnectionException
+from ..registers.read_planner import group_reads
 from ..register_defs_cache import get_register_definitions
 from ..registers.register_def import RegisterDef
 from ..scanner import (
