+++ custom_components/thessla_green_modbus/modbus/__init__.py
@@ -1,0 +1,1 @@
+"""Modbus communication helpers."""
