+++ custom_components/thessla_green_modbus/modbus_helpers.py
@@ -1,646 +1,70 @@
-"""Utility helpers for Modbus communication and read grouping."""
+"""Compatibility re-exports - functions have moved to focused modules.
+
+This module is kept for backwards compatibility. All functions are now
+implemented in focused modules under ``modbus/`` and ``registers/``.
+"""
 
 from __future__ import annotations
 
-import asyncio
-import inspect
-import logging
-import random
-import weakref
-from collections.abc import Awaitable, Callable, Iterable
-from dataclasses import dataclass
-from typing import Any
-
-from . import const
-from .modbus_exceptions import ModbusIOException
-
-_LOGGER = logging.getLogger(__name__)
-
-# Cache which keyword variant a given function accepts
-_KWARG_CACHE: weakref.WeakKeyDictionary[Callable[..., Awaitable[Any]], str | None] = (
-    weakref.WeakKeyDictionary()
+from .modbus.call import (
+    _KWARG_CACHE,
+    _PreparedCall,
+    _SIG_CACHE,
+    _apply_attempt_delay,
+    _calculate_backoff_delay,
+    _calculate_batch_size,
+    _call_modbus,
+    _classify_modbus_exception,
+    _dispatch_modbus_call,
+    _get_signature,
+    _invoke_with_slave_kwarg,
+    _normalize_positional_and_keyword_args,
+    _prepare_modbus_call,
+    _raise_mapped_call_exception,
+    _resolve_slave_kwarg,
+    _should_apply_external_timeout,
+    async_maybe_await,
 )
-# Cache function signatures to avoid repeated inspection
-_SIG_CACHE: weakref.WeakKeyDictionary[Callable[..., Awaitable[Any]], inspect.Signature] = (
-    weakref.WeakKeyDictionary()
+from .modbus.client_close import async_close_client, async_maybe_await_close
+from .modbus.frame_logging import (
+    _build_request_frame,
+    _log_modbus_request,
+    _log_modbus_response,
+    _mask_frame,
+)
+from .modbus.framer import get_rtu_framer
+from .registers.read_planner import (
+    chunk_register_range,
+    chunk_register_values,
+    group_reads,
 )
 
-
-def _get_signature(func: Callable[..., Awaitable[Any]]) -> inspect.Signature | None:
-    """Return a cached signature when introspection is available."""
-
-    signature = _SIG_CACHE.get(func)
-    if signature is not None:
-        return signature
-
-    try:
-        signature = inspect.signature(func)
-    except (TypeError, ValueError):
-        return None
-
-    _SIG_CACHE[func] = signature
-    return signature
-
-
-def _should_apply_external_timeout(func: Callable[..., Awaitable[Any]]) -> bool:
-    """Return whether ``asyncio.wait_for`` should wrap ``func`` calls.
-
-    Pymodbus clients already implement their own timeout and cancellation
-    handling. Wrapping those calls in an external ``wait_for`` can cancel
-    requests mid-flight, leaving stale transaction IDs and causing follow-up
-    "request ask for transaction_id" desynchronization errors.
-    """
-
-    module_name = getattr(func, "__module__", "") or ""
-    return not module_name.startswith("pymodbus")
-
-
-async def async_maybe_await(result: Any) -> Any:
-    """Await a result only if it is awaitable."""
-
-    if inspect.isawaitable(result):
-        return await result
-    return result
-
-
-try:
-    from pymodbus.framer import FramerType
-except (ImportError, ModuleNotFoundError):  # pragma: no cover
-    FramerType = None
-
-try:
-    from pymodbus.framer import ModbusRtuFramer
-except (ImportError, ModuleNotFoundError):  # pragma: no cover
-    try:
-        from pymodbus.framer.rtu_framer import ModbusRtuFramer
-    except (ImportError, ModuleNotFoundError):  # pragma: no cover
-        ModbusRtuFramer = None
-
-
-def get_rtu_framer() -> Any | None:
-    """Return a Modbus RTU framer class/enum when available."""
-
-    if FramerType is not None:
-        try:
-            return FramerType.RTU
-        except (AttributeError, ValueError):  # pragma: no cover - unexpected
-            return None
-    if ModbusRtuFramer is not None:
-        return ModbusRtuFramer
-    return None
-
-
-async def async_maybe_await_close(obj: Any | None) -> None:
-    """Close a Modbus client safely across sync/async implementations."""
-
-    if obj is None:
-        return
-
-    close = getattr(obj, "close", None)
-    if close is None or not callable(close):
-        return
-
-    try:
-        result = close()
-    except (
-        AttributeError,
-        OSError,
-        RuntimeError,
-        TypeError,
-        ValueError,
-    ) as exc:  # pragma: no cover - defensive
-        _LOGGER.debug("Error calling close on Modbus client: %s", exc)
-        return
-
-    try:
-        await async_maybe_await(result)
-    except (
-        AttributeError,
-        OSError,
-        RuntimeError,
-        TypeError,
-        ValueError,
-    ) as exc:  # pragma: no cover - defensive
-        _LOGGER.debug("Error awaiting Modbus client close: %s", exc)
-
-
-async def async_close_client(client: Any | None) -> None:
-    """Wrapper for safe Modbus client close."""
-
-    await async_maybe_await_close(client)
-
-
-def _mask_frame(frame: bytes) -> str:
-    """Return a hex representation of ``frame`` with the slave ID masked."""
-
-    if not frame:
-        return ""
-
-    hex_str = frame.hex()
-    if len(hex_str) >= 2:
-        return f"**{hex_str[2:]}"
-    return hex_str
-
-
-_READ_FC: dict[str, int] = {
-    "read_input_registers": 4,
-    "read_holding_registers": 3,
-    "read_coils": 1,
-    "read_discrete_inputs": 2,
-}
-
-
-def _encode_read_frame(
-    slave_id: int, fc: int, positional: list[Any], kwargs: dict[str, Any]
-) -> bytes:
-    """Encode a standard Modbus read request frame (addr + count)."""
-    addr = int(positional[0])
-    count = int(kwargs.get("count", positional[1] if len(positional) > 1 else 1))
-    return bytes([slave_id, fc, addr >> 8, addr & 255, count >> 8, count & 255])
-
-
-def _build_request_frame(
-    func_name: str, slave_id: int, positional: list[Any], kwargs: dict[str, Any]
-) -> bytes:
-    """Best-effort Modbus request frame builder for logging."""
-
-    try:
-        if func_name in _READ_FC:
-            return _encode_read_frame(slave_id, _READ_FC[func_name], positional, kwargs)
-        if func_name == "write_register":
-            addr = int(kwargs.get("address", positional[0]))
-            value = int(kwargs.get("value", positional[1] if len(positional) > 1 else 0))
-            return bytes([slave_id, 6, addr >> 8, addr & 255, value >> 8, value & 255])
-        if func_name == "write_registers":
-            addr = int(kwargs.get("address", positional[0]))
-            values = [
-                int(v) for v in kwargs.get("values", positional[1] if len(positional) > 1 else [])
-            ]
-            qty = len(values)
-            frame = bytearray(
-                [
-                    slave_id,
-                    16,
-                    addr >> 8,
-                    addr & 255,
-                    qty >> 8,
-                    qty & 255,
-                    qty * 2,
-                ]
-            )
-            for v in values:
-                frame.extend([v >> 8, v & 255])
-            return bytes(frame)
-        if func_name == "write_coil":
-            addr = int(kwargs.get("address", positional[0]))
-            value = (
-                65280 if kwargs.get("value", positional[1] if len(positional) > 1 else False) else 0
-            )
-            return bytes([slave_id, 5, addr >> 8, addr & 255, value >> 8, value & 255])
-    except (ValueError, TypeError, IndexError) as err:
-        _LOGGER.debug("Failed to build request frame: %s", err)
-        return b""
-    return b""
-
-
-def _calculate_backoff_delay(
-    *,
-    base: float,
-    attempt: int,
-    jitter: float | tuple[float, float] | None,
-) -> float:
-    """Return the delay for the given ``attempt`` including optional jitter."""
-
-    if base <= 0 or attempt <= 1:
-        return 0.0
-
-    delay = float(base) * (2 ** (attempt - 2))
-
-    if jitter:
-        if isinstance(jitter, int | float):
-            jitter_min = 0.0
-            jitter_max = float(jitter)
-        else:
-            jitter_min, jitter_max = (float(jitter[0]), float(jitter[1]))
-        if jitter_max < jitter_min:
-            jitter_min, jitter_max = jitter_max, jitter_min
-        delay += random.uniform(jitter_min, jitter_max)
-
-    return float(max(delay, 0.0))
-
-
-def _normalize_positional_and_keyword_args(
-    signature: inspect.Signature | None,
-    args: tuple[Any, ...],
-    kwargs: dict[str, Any],
-) -> tuple[list[Any], dict[str, inspect.Parameter]]:
-    """Move positional values targeting keyword-only params into ``kwargs``."""
-    params: dict[str, inspect.Parameter] = (
-        dict(signature.parameters) if signature is not None else {}
-    )
-    positional: list[Any] = []
-    if signature is None:
-        return list(args), params
-
-    param_iter = iter(params.values())
-    for arg in args:
-        try:
-            param = next(param_iter)
-        except StopIteration:
-            positional.append(arg)
-            continue
-
-        if param.kind is inspect.Parameter.KEYWORD_ONLY:
-            kwargs[param.name] = arg
-        else:
-            positional.append(arg)
-    return positional, params
-
-
-def _resolve_slave_kwarg(
-    func: Callable[..., Awaitable[Any]],
-    params: dict[str, inspect.Parameter],
-    signature: inspect.Signature | None,
-) -> str:
-    """Return cached keyword variant supported by Modbus client callable."""
-    kwarg = _KWARG_CACHE.get(func)
-    if kwarg is not None:
-        return kwarg
-
-    if "device_id" in params and params["device_id"].kind is not inspect.Parameter.POSITIONAL_ONLY:
-        kwarg = "device_id"
-    elif "slave" in params and params["slave"].kind is not inspect.Parameter.POSITIONAL_ONLY:
-        kwarg = "slave"
-    elif "unit" in params and params["unit"].kind is not inspect.Parameter.POSITIONAL_ONLY:
-        kwarg = "unit"
-    else:
-        kwarg = "slave" if signature is None else ""
-
-    _KWARG_CACHE[func] = kwarg
-    return kwarg
-
-
-async def _invoke_with_slave_kwarg(
-    func: Callable[..., Awaitable[Any]],
-    positional: list[Any],
-    kwargs: dict[str, Any],
-    kwarg: str,
-    slave_id: int,
-) -> Any:
-    """Invoke wrapped callable using detected slave keyword convention."""
-    if kwarg == "device_id":
-        return await async_maybe_await(func(*positional, device_id=slave_id, **kwargs))
-    if kwarg == "slave":
-        return await async_maybe_await(func(*positional, slave=slave_id, **kwargs))
-    if kwarg == "unit":
-        return await async_maybe_await(func(*positional, unit=slave_id, **kwargs))
-    return await async_maybe_await(func(*positional, **kwargs))
-
-
-def _log_modbus_request(
-    *,
-    func_name: str,
-    slave_id: int,
-    positional: list[Any],
-    kwargs: dict[str, Any],
-) -> None:
-    """Emit request diagnostics at debug level."""
-    request_frame = _build_request_frame(func_name, slave_id, positional, kwargs)
-    if request_frame:
-        _LOGGER.debug("Modbus request: %s", _mask_frame(request_frame))
-        return
-    if _LOGGER.isEnabledFor(logging.DEBUG):
-        _LOGGER.debug(
-            "Sending %s to slave %s: args=%s kwargs=%s", func_name, slave_id, positional, kwargs
-        )
-
-
-def _log_modbus_response(func_name: str, response: Any) -> None:
-    """Emit response diagnostics at debug level."""
-    if _LOGGER.isEnabledFor(logging.DEBUG):
-        try:
-            encoded = response.encode() if hasattr(response, "encode") else b""
-        except (AttributeError, ValueError, TypeError, UnicodeError) as err:
-            _LOGGER.debug("Failed to encode Modbus response: %s", err)
-            encoded = b""
-        except (OSError, RuntimeError) as err:  # pragma: no cover - unexpected
-            _LOGGER.exception("Unexpected error encoding Modbus response: %s", err)
-            encoded = b""
-        if encoded:
-            _LOGGER.debug("Modbus response: %s", _mask_frame(encoded))
-        else:
-            _LOGGER.debug("Received from %s: %s", func_name, response)
-        return
-
-    try:
-        encoded = response.encode() if hasattr(response, "encode") else b""
-    except (AttributeError, OSError, RuntimeError, TypeError, ValueError):
-        encoded = b""
-    if encoded:
-        _LOGGER.debug("Modbus response: %s", _mask_frame(encoded))
-
-
-def _calculate_batch_size(kwargs: dict[str, Any]) -> int:
-    """Calculate batch size metadata for request logging."""
-    return kwargs.get("count") or len(kwargs.get("values", [])) or 1
-
-
-@dataclass(frozen=True, slots=True)
-class _PreparedCall:
-    positional: list[Any]
-    kwarg: str
-    func_name: str
-    batch_size: int
-    delay: float
-
-
-def _prepare_modbus_call(
-    func: Callable[..., Awaitable[Any]],
-    args: tuple[Any, ...],
-    kwargs: dict[str, Any],
-    *,
-    attempt: int,
-    backoff: float,
-    backoff_jitter: float | tuple[float, float] | None,
-    apply_backoff: bool,
-) -> tuple[list[Any], str, str, int, float]:
-    """Prepare call metadata for ``_call_modbus`` without mutating behavior."""
-
-    signature = _get_signature(func)
-    positional, params = _normalize_positional_and_keyword_args(signature, args, kwargs)
-    kwarg = _resolve_slave_kwarg(func, params, signature)
-    func_name = getattr(func, "__name__", repr(func))
-    batch_size = _calculate_batch_size(kwargs)
-    delay = (
-        _calculate_backoff_delay(base=backoff, attempt=attempt, jitter=backoff_jitter)
-        if apply_backoff
-        else 0.0
-    )
-    return positional, kwarg, func_name, batch_size, delay
-
-
-def _classify_modbus_exception(err: Exception) -> str:
-    """Return exception class for Modbus call logging."""
-    if isinstance(err, ModbusIOException) and "request cancelled" in str(err).lower():
-        return "cancelled"
-    return "failed"
-
-
-def _log_call_attempt(
-    prepared: _PreparedCall,
-    *,
-    slave_id: int,
-    attempt: int,
-    max_attempts: int,
-    kwargs: dict[str, Any],
-) -> None:
-    """Emit pre-dispatch attempt diagnostics at debug level."""
-    _LOGGER.debug(
-        "Calling %s on slave %s (batch=%s attempt %s/%s)",
-        prepared.func_name,
-        slave_id,
-        prepared.batch_size,
-        attempt,
-        max_attempts,
-    )
-    _log_modbus_request(
-        func_name=prepared.func_name,
-        slave_id=slave_id,
-        positional=prepared.positional,
-        kwargs=kwargs,
-    )
-
-
-async def _dispatch_modbus_call(
-    func: Callable[..., Awaitable[Any]],
-    positional: list[Any],
-    kwargs: dict[str, Any],
-    kwarg: str,
-    slave_id: int,
-    timeout: float | None,
-) -> Any:
-    """Dispatch a Modbus callable with optional external timeout."""
-
-    async def _invoke() -> Any:
-        return await _invoke_with_slave_kwarg(func, positional, kwargs, kwarg, slave_id)
-
-    if timeout is not None and _should_apply_external_timeout(func):
-        return await asyncio.wait_for(_invoke(), timeout=timeout)
-    return await _invoke()
-
-
-async def _apply_attempt_delay(
-    *,
-    delay: float,
-    func_name: str,
-    attempt: int,
-    max_attempts: int,
-) -> None:
-    """Apply pre-attempt delay while preserving cancellation diagnostics."""
-    if delay <= 0:
-        return
-
-    _LOGGER.debug(
-        "Delaying %.3fs before attempt %s/%s of %s",
-        delay,
-        attempt,
-        max_attempts,
-        func_name,
-    )
-    try:
-        await asyncio.sleep(delay)
-    except asyncio.CancelledError:
-        _LOGGER.debug(
-            "Sleep cancelled before calling %s attempt %s/%s",
-            func_name,
-            attempt,
-            max_attempts,
-        )
-        raise
-
-
-def _raise_mapped_call_exception(
-    err: Exception,
-    *,
-    func_name: str,
-    attempt: int,
-    max_attempts: int,
-) -> None:
-    """Log call classification and re-raise preserving existing behavior."""
-    if isinstance(err, TimeoutError):
-        _LOGGER.debug("Call to %s timed out on attempt %s/%s", func_name, attempt, max_attempts)
-        raise TimeoutError("Modbus request timed out") from err
-
-    classification = _classify_modbus_exception(err)
-    if classification == "cancelled":
-        _LOGGER.debug("Call to %s cancelled on attempt %s/%s", func_name, attempt, max_attempts)
-    else:
-        _LOGGER.debug("Call to %s failed on attempt %s/%s", func_name, attempt, max_attempts)
-    raise
-
-
-async def _call_modbus(
-    func: Callable[..., Awaitable[Any]],
-    slave_id: int,
-    *args: Any,
-    attempt: int = 1,
-    max_attempts: int = 1,
-    timeout: float | None = None,
-    backoff: float = 0.0,
-    backoff_jitter: float | tuple[float, float] | None = None,
-    apply_backoff: bool = True,
-    **kwargs: Any,
-) -> Any:
-    """Invoke a Modbus function handling Modbus client keyword variants.
-
-    The function signature is inspected to determine whether the wrapped
-    callable expects a ``device_id`` or ``slave`` keyword argument.  If neither is
-    present the function is called without either keyword.  The chosen keyword
-    (or lack thereof) is cached per callable for subsequent invocations.
-    """
-
-    positional, kwarg, func_name, batch_size, delay = _prepare_modbus_call(
-        func,
-        args,
-        kwargs,
-        attempt=attempt,
-        backoff=backoff,
-        backoff_jitter=backoff_jitter,
-        apply_backoff=apply_backoff,
-    )
-
-    prepared = _PreparedCall(positional, kwarg, func_name, batch_size, delay)
-    await _apply_attempt_delay(
-        delay=prepared.delay,
-        func_name=prepared.func_name,
-        attempt=attempt,
-        max_attempts=max_attempts,
-    )
-
-    _log_call_attempt(
-        prepared,
-        slave_id=slave_id,
-        attempt=attempt,
-        max_attempts=max_attempts,
-        kwargs=kwargs,
-    )
-
-    try:
-        response = await _dispatch_modbus_call(
-            func=func,
-            positional=prepared.positional,
-            kwargs=kwargs,
-            kwarg=prepared.kwarg,
-            slave_id=slave_id,
-            timeout=timeout,
-        )
-    except TimeoutError as err:
-        _raise_mapped_call_exception(
-            err, func_name=prepared.func_name, attempt=attempt, max_attempts=max_attempts
-        )
-    except asyncio.CancelledError:
-        _LOGGER.debug(
-            "Call to %s cancelled on attempt %s/%s", prepared.func_name, attempt, max_attempts
-        )
-        raise
-    except (
-        AttributeError,
-        ModbusIOException,
-        OSError,
-        RuntimeError,
-        TypeError,
-        ValueError,
-    ) as err:
-        _raise_mapped_call_exception(
-            err, func_name=prepared.func_name, attempt=attempt, max_attempts=max_attempts
-        )
-
-    _log_modbus_response(prepared.func_name, response)
-    return response
-
-
-def group_reads(
-    addresses: Iterable[int],
-    max_block_size: int | None = None,
-    boundaries: frozenset[int] | None = None,
-) -> list[tuple[int, int]]:
-    """Group raw register addresses into contiguous read blocks.
-
-    The addresses are sorted and sequential ranges are merged up to
-    ``max_block_size`` entries.  The returned list contains ``(start, length)``
-    tuples suitable for bulk Modbus read operations.
-
-    ``boundaries`` is an optional set of addresses at which a new batch must
-    start regardless of contiguity.  Use this to prevent batches from spanning
-    register regions that a device firmware cannot handle in a single request
-    (e.g. AirPack4 FW 3.11 rejects FC03 reads that cross addr 15→16).
-    """
-
-    if max_block_size is None:
-        max_block_size = const.MAX_REGS_PER_REQUEST
-    max_block_size = min(max_block_size, const.MAX_REGS_PER_REQUEST)
-    sorted_addresses = sorted(set(addresses))
-    if not sorted_addresses:
-        return []
-
-    groups: list[tuple[int, int]] = []
-    start = prev = sorted_addresses[0]
-    for addr in sorted_addresses[1:]:
-        force_split = boundaries is not None and addr in boundaries
-        if not force_split and addr == prev + 1 and (addr - start + 1) <= max_block_size:
-            prev = addr
-            continue
-        groups.append((start, prev - start + 1))
-        start = prev = addr
-    groups.append((start, prev - start + 1))
-    return groups
-
-
-def chunk_register_range(
-    start: int,
-    count: int,
-    max_block_size: int | None = None,
-) -> list[tuple[int, int]]:
-    """Split a contiguous register range into safe-sized chunks."""
-
-    if count <= 0:
-        return []
-
-    if max_block_size is None:
-        max_block_size = const.MAX_REGS_PER_REQUEST
-    max_block_size = min(max_block_size, const.MAX_REGS_PER_REQUEST)
-    if max_block_size < 1:
-        max_block_size = 1
-
-    chunks: list[tuple[int, int]] = []
-    remaining = count
-    current = start
-    while remaining > 0:
-        chunk_len = min(remaining, max_block_size)
-        chunks.append((current, chunk_len))
-        current += chunk_len
-        remaining -= chunk_len
-    return chunks
-
-
-def chunk_register_values(
-    start: int,
-    values: list[int],
-    max_block_size: int | None = None,
-) -> list[tuple[int, list[int]]]:
-    """Split register values into safe-sized chunks with updated addresses."""
-
-    if not values:
-        return []
-
-    if max_block_size is None:
-        max_block_size = const.MAX_REGS_PER_REQUEST
-    max_block_size = min(max_block_size, const.MAX_REGS_PER_REQUEST)
-    if max_block_size < 1:
-        max_block_size = 1
-
-    chunks: list[tuple[int, list[int]]] = []
-    for offset in range(0, len(values), max_block_size):
-        chunk = values[offset : offset + max_block_size]
-        chunks.append((start + offset, chunk))
-    return chunks
+__all__ = [
+    "_KWARG_CACHE",
+    "_PreparedCall",
+    "_SIG_CACHE",
+    "_apply_attempt_delay",
+    "_build_request_frame",
+    "_calculate_backoff_delay",
+    "_calculate_batch_size",
+    "_call_modbus",
+    "_classify_modbus_exception",
+    "_dispatch_modbus_call",
+    "_get_signature",
+    "_invoke_with_slave_kwarg",
+    "_log_modbus_request",
+    "_log_modbus_response",
+    "_mask_frame",
+    "_normalize_positional_and_keyword_args",
+    "_prepare_modbus_call",
+    "_raise_mapped_call_exception",
+    "_resolve_slave_kwarg",
+    "_should_apply_external_timeout",
+    "async_close_client",
+    "async_maybe_await",
+    "async_maybe_await_close",
+    "chunk_register_range",
+    "chunk_register_values",
+    "get_rtu_framer",
+    "group_reads",
+]
