+++ custom_components/thessla_green_modbus/modbus_transport.py
@@ -1,6 +1,6 @@
 """Transport abstractions for Modbus communication."""
 
-from .modbus_helpers import get_rtu_framer
+from .modbus.framer import get_rtu_framer
 from .modbus_transport_base import classify_transport_error
 from .transport.base import BaseModbusTransport
 from .transport.crc import append_crc as _append_crc
