+++ custom_components/thessla_green_modbus/registers/read_planner.py
@@ -3,10 +3,98 @@
 from __future__ import annotations
 
 from collections.abc import Callable, Iterable
+from typing import Any
 
-from ..modbus_helpers import group_reads
+from .. import const
 from .definition import ReadPlan
 from .register_def import RegisterDef
+
+
+def group_reads(
+    addresses: Iterable[int],
+    max_block_size: int | None = None,
+    boundaries: frozenset[int] | None = None,
+) -> list[tuple[int, int]]:
+    """Group raw register addresses into contiguous read blocks.
+
+    The addresses are sorted and sequential ranges are merged up to
+    ``max_block_size`` entries.  The returned list contains ``(start, length)``
+    tuples suitable for bulk Modbus read operations.
+
+    ``boundaries`` is an optional set of addresses at which a new batch must
+    start regardless of contiguity.  Use this to prevent batches from spanning
+    register regions that a device firmware cannot handle in a single request
+    (e.g. AirPack4 FW 3.11 rejects FC03 reads that cross addr 15→16).
+    """
+
+    if max_block_size is None:
+        max_block_size = const.MAX_REGS_PER_REQUEST
+    max_block_size = min(max_block_size, const.MAX_REGS_PER_REQUEST)
+    sorted_addresses = sorted(set(addresses))
+    if not sorted_addresses:
+        return []
+
+    groups: list[tuple[int, int]] = []
+    start = prev = sorted_addresses[0]
+    for addr in sorted_addresses[1:]:
+        force_split = boundaries is not None and addr in boundaries
+        if not force_split and addr == prev + 1 and (addr - start + 1) <= max_block_size:
+            prev = addr
+            continue
+        groups.append((start, prev - start + 1))
+        start = prev = addr
+    groups.append((start, prev - start + 1))
+    return groups
+
+
+def chunk_register_range(
+    start: int,
+    count: int,
+    max_block_size: int | None = None,
+) -> list[tuple[int, int]]:
+    """Split a contiguous register range into safe-sized chunks."""
+
+    if count <= 0:
+        return []
+
+    if max_block_size is None:
+        max_block_size = const.MAX_REGS_PER_REQUEST
+    max_block_size = min(max_block_size, const.MAX_REGS_PER_REQUEST)
+    if max_block_size < 1:
+        max_block_size = 1
+
+    chunks: list[tuple[int, int]] = []
+    remaining = count
+    current = start
+    while remaining > 0:
+        chunk_len = min(remaining, max_block_size)
+        chunks.append((current, chunk_len))
+        current += chunk_len
+        remaining -= chunk_len
+    return chunks
+
+
+def chunk_register_values(
+    start: int,
+    values: list[int],
+    max_block_size: int | None = None,
+) -> list[tuple[int, list[int]]]:
+    """Split register values into safe-sized chunks with updated addresses."""
+
+    if not values:
+        return []
+
+    if max_block_size is None:
+        max_block_size = const.MAX_REGS_PER_REQUEST
+    max_block_size = min(max_block_size, const.MAX_REGS_PER_REQUEST)
+    if max_block_size < 1:
+        max_block_size = 1
+
+    chunks: list[tuple[int, list[int]]] = []
+    for offset in range(0, len(values), max_block_size):
+        chunk = values[offset : offset + max_block_size]
+        chunks.append((start + offset, chunk))
+    return chunks
 
 
 def plan_group_reads(
