+++ custom_components/thessla_green_modbus/scanner/core.py
@@ -22,8 +22,8 @@ from ..const import (
 from ..const import (
     SERIAL_STOP_BITS_MAP as _SERIAL_STOP_BITS_MAP,
 )
-from ..modbus_helpers import async_maybe_await_close
-from ..modbus_helpers import group_reads as _group_reads
+from ..modbus.client_close import async_maybe_await_close
+from ..registers.read_planner import group_reads as _group_reads
 from ..registers.loader import (
     async_get_all_registers,
 )
