+++ custom_components/thessla_green_modbus/scanner/io_read.py
@@ -9,6 +9,7 @@ from typing import Any
 from pymodbus.client import AsyncModbusTcpClient
 
 from ..modbus_exceptions import ConnectionException, ModbusException, ModbusIOException
+from ..registers.read_planner import chunk_register_range
 from ..transport.retry import ErrorKind, classify_transport_error
 from .io_core import (
     _call_modbus_with_fallback,
