+++ custom_components/thessla_green_modbus/scanner/orchestration.py
@@ -17,7 +17,7 @@ from ..const import (
     SERIAL_STOP_BITS_MAP,
 )
 from ..modbus_exceptions import ConnectionException, ModbusException, ModbusIOException
-from ..modbus_helpers import group_reads as _group_reads
+from ..registers.read_planner import group_reads as _group_reads
 from ..scanner_device_info import ScannerDeviceInfo
 from ..transport.rtu import RtuModbusTransport
 from . import custom_scan as scanner_custom_scan
